Split ifaddr related functions out into a separate file
